-Wextra clean

+++ libterm/tputs.c
@@ -116,7 +116,7 @@ tputs(const char *cp, int affcnt, int (*outc)(int))
 	 */
 	if (i == 0)
 		return 1;
-	if (ospeed <= 0 || ospeed >= (sizeof tmspc10 / sizeof tmspc10[0]))
+	if (ospeed <= 0 || ospeed >= (int)(sizeof tmspc10 / sizeof tmspc10[0]))
 		return 1;
 
 	/*
